move tests to their own place
